move `calculateSeasonScores` to `scores.ts`
