+++ lib/leaderboard/leaderboard_client.ts
@@ -1,6 +1,3 @@
-import { DAY } from "lc-dailies/deps.ts";
-import type * as lc from "lc-dailies/lib/lc/mod.ts";
-
 /**
  * Player is a registered player in the leaderboard.
  */
@@ -32,6 +29,36 @@ export interface Submission {
 }
 
 /**
+ * LCQuestion is a Leetcode question.
+ */
+export interface LCQuestion {
+  /**
+   * name is the name of the daily question.
+   */
+  name: string;
+
+  /**
+   * date is the date the daily question was posted in the format of YYYY-MM-DD.
+   */
+  date: string;
+
+  /**
+   * title is the title of the daily question.
+   */
+  title: string;
+
+  /**
+   * difficulty is the difficulty of the daily question.
+   */
+  difficulty: string;
+
+  /**
+   * url is the link of the daily question.
+   */
+  url: string;
+}
+
+/**
  * Season is a season of the leaderboard.
  */
 export interface Season {
@@ -53,7 +80,7 @@ export interface Season {
   /**
    * questions is the map of questions in the season.
    */
-  questions: { [lc_question_name: string]: lc.DailyQuestion };
+  questions: { [lc_question_name: string]: LCQuestion };
 
   /**
    * submissions is the map of submissions in the season.
@@ -121,109 +148,4 @@ export interface LeaderboardClient {
    * listSeasons gets a list of season IDs of the leaderboard.
    */
   listSeasons(): Promise<Season[]>;
-}
-
-/**
- * CalculateScoresOptions is the options for calculateScores.
- */
-export interface CalculateScoresOptions {
-  /**
-   * season is the season to calculate the score for.
-   */
-  season: Season;
-
-  /**
-   * possibleHighestScore is the highest possible score.
-   */
-  possibleHighestScore: number;
-
-  /**
-   * possibleLowestScore is the lowest possible score.
-   */
-  possibleLowestScore: number;
-
-  /**
-   * duration is the amount of milliseconds it takes to linearly interpolate
-   * between the highest and lowest possible scores.
-   */
-  duration: number;
-}
-
-/**
- * calculateSubmissionScore calculates the score of a submission.
- */
-export function calculateSubmissionScore(
-  submission: Submission,
-  question: lc.DailyQuestion,
-  options: CalculateScoresOptions,
-): number {
-  const questionDate = new Date(`${question.date} GMT`);
-  const submissionDate = new Date(submission.date);
-  const msElapsed = submissionDate.getTime() - questionDate.getTime();
-  const ratio = Math.min(Math.max(msElapsed / options.duration, 0), 1);
-  const questionScore =
-    ((options.possibleHighestScore - options.possibleLowestScore) *
-      ratio) + options.possibleLowestScore;
-  return Math.ceil(questionScore);
-}
-
-/**
- * calculatePlayerScore calculates the score of a player in a season.
- *
- * See:
- * https://github.com/acmcsufoss/lc-dailies/issues/32#issuecomment-1728904942
- */
-export function calculatePlayerScore(
-  playerID: string,
-  options: CalculateScoresOptions,
-): number {
-  const submissions = options.season.submissions[playerID];
-  if (!submissions) {
-    return 0;
-  }
-
-  const scoreSum = Object.entries(submissions)
-    .reduce((score, [questionID, submission]) => {
-      const question = options.season.questions[questionID];
-      if (!question) {
-        return score;
-      }
-
-      return score + calculateSubmissionScore(
-        submission,
-        question,
-        options,
-      );
-    }, 0);
-
-  return scoreSum;
-}
-
-/**
- * calculateSeasonScores calculates the scores of all players in a season.
- *
- * Returns a map of player ID to score.
- */
-export function calculateSeasonScores(
-  options: CalculateScoresOptions,
-): Record<string, number> {
-  return Object.keys(options.season.players)
-    .reduce((scores, playerID) => {
-      scores[playerID] = calculatePlayerScore(playerID, options);
-      return scores;
-    }, {} as Record<string, number>);
-}
-
-/**
- * makeDefaultCalculateScoresOptions creates a default CalculateScoresOptions.
- */
-export function makeDefaultCalculateScoresOptions(
-  season: Season,
-): CalculateScoresOptions {
-  return {
-    season,
-    possibleHighestScore: 100,
-    possibleLowestScore: 50,
-    duration: DAY,
-  };
 }
