+++ lib/leaderboard/mod.ts
@@ -1,1 +1,2 @@
 export * from "./leaderboard_client.ts";
+export * from "./scores.ts";
